refactor: Moved the getScope eventbus functions to their own file
